move everything out of init; fix from imports and some typing errors
