+++ worlds/stardew_valley/test/TestBundles.py
@@ -1,6 +1,6 @@
 import unittest
 
-from . import SVTestBase
+from .bases import SVTestBase
 from .. import BundleRandomization
 from ..data.bundle_data import all_bundle_items_except_money, quality_crops_items_thematic, quality_foraging_items, quality_fish_items
 from ..options import BundlePlando
@@ -87,4 +87,3 @@ class TestRemixedAnywhereBundles(SVTestBase):
         for bundle_name in self.fish_bundle_names:
             with self.subTest(f"{bundle_name}"):
                 self.assertIn(bundle_name, location_names)
-
