+++ worlds/stardew_valley/test/TestFill.py
@@ -1,5 +1,5 @@
-from . import SVTestBase
 from .assertion import WorldAssertMixin
+from .bases import SVTestBase
 from .options.presets import minimal_locations_maximal_items
 from .. import options
 from ..mods.mod_data import ModNames
