+++ worlds/stardew_valley/test/TestNumberLocations.py
@@ -1,4 +1,4 @@
-from . import SVTestBase
+from .bases import SVTestBase
 from .options.presets import default_6_x_x, allsanity_no_mods_6_x_x, allsanity_mods_6_x_x_exclude_disabled, get_minsanity_options, \
     minimal_locations_maximal_items, minimal_locations_maximal_items_with_island
 from .. import location_table
