+++ worlds/stardew_valley/test/long/TestOptionsLong.py
@@ -4,8 +4,8 @@ from typing import ClassVar
 
 from BaseClasses import get_seed
 from test.param import classvar_matrix
-from .. import SVTestCase, solo_multiworld, skip_long_tests
 from ..assertion.world_assert import WorldAssertMixin
+from ..bases import skip_long_tests, SVTestCase, solo_multiworld
 from ..options.option_names import all_option_choices
 from ... import options
