+++ worlds/stardew_valley/test/rules/TestDonations.py
@@ -1,8 +1,8 @@
+from ..bases import SVTestBase
 from ... import options
 from ...locations import locations_by_tag, LocationTags, location_table
 from ...strings.entrance_names import Entrance
 from ...strings.region_names import Region
-from ...test import SVTestBase
 
 
 class TestDonationLogicAll(SVTestBase):
